src/test/java/com/github/artemkorsakov/figure/TriangleTests.java
package com.github.artemkorsakov.figure;

import org.testng.Assert;
import org.testng.annotations.Test;

import java.awt.*;

public class TriangleTests {
    @Test
    public void test() {
        Assert.assertEquals(new Triangle(new Point(-340, 495), new Point(-153, -910), new Point(835, -947)).getPointOnTriangleType(new Point(0, 0)), Triangle.PointOnTriangleType.INSIDE);
        Assert.assertEquals(new Triangle(new Point(-175, 41), new Point(-421, -714), new Point(574, -645)).getPointOnTriangleType(new Point(0, 0)), Triangle.PointOnTriangleType.OUTSIDE);
    }

    @Test
    public void testIsZeroPointInside() {
        Assert.assertTrue(new Triangle(new Point(-340, 495), new Point(-153, -910), new Point(835, -947)).isZeroPointInside());
        Assert.assertFalse(new Triangle(new Point(-175, 41), new Point(-421, -714), new Point(574, -645)).isZeroPointInside());
    }
}
